apps/api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
sqlalchemy==2.0.23
oracledb==2.0.0
aiofiles==23.2.1
python-multipart==0.0.6
google-genai
